# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenNotebook",
    "description": "Collections of sources, and the decks, audio overviews, mind maps and notes made from them. Everything the web app does goes through this API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "List Collections",
        "description": "Your collections, most recently updated first, with counts of what\neach holds, a page at a time: `X-Next-Offset` says where the next page\nstarts when there is one.",
        "operationId": "list_collections",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "The most collections in one answer; 200 unless set",
              "default": 200,
              "title": "Limit"
            },
            "description": "The most collections in one answer; 200 unless set"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "description": "Where this page starts: the `X-Next-Offset` header of the page before",
              "default": 0,
              "title": "Offset"
            },
            "description": "Where this page starts: the `X-Next-Offset` header of the page before"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionSummary"
                  },
                  "title": "Response List Collections"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Create Collection",
        "description": "Start a collection: an empty set of sources to add to and build outputs from.",
        "operationId": "create_collection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collections/{cid}": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Get Collection",
        "description": "One collection, with its decks and audio overviews.",
        "operationId": "get_collection",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "collections"
        ],
        "summary": "Update Collection",
        "description": "Rename or pin a collection. A read-only copy can be pinned, not renamed.",
        "operationId": "update_collection",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Delete Collection",
        "description": "Delete a collection with its sources and everything made from them.",
        "operationId": "delete_collection",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collections/{cid}/cover": {
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Refresh Cover",
        "description": "Design the collection's cover again, from what it holds now, and answer\nonce it is drawn. With covers off in Settings no model is asked and the\ncover stays the one drawn from the title. If the model fails, the cover\nthe collection had is kept and the answer says why.",
        "operationId": "refresh_cover",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Read Cover",
        "description": "The collection's cover: a self-contained HTML page (an SVG, no script,\nnothing fetched) for a sandboxed iframe. Cached for good when `v` is the\nversion drawn now, so a new version is a new URL.",
        "operationId": "read_cover",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          },
          {
            "name": "v",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "The `cover_version` the page was listed with",
              "default": "",
              "title": "V"
            },
            "description": "The `cover_version` the page was listed with"
          },
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "`light`, else the dark default",
              "default": "dark",
              "title": "Theme"
            },
            "description": "`light`, else the dark default"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collections/{cid}/events": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Collection Events",
        "description": "Server-sent events while a collection is open: `collection` (its\nsummary, as the list has it), `outputs` (its decks and audio overviews),\n`progress` (`session_id`, `step`, `steps_done`, `steps_total` of one being\nmade), `sources` (its sources), each when the stream starts and again when\nit changes, and `gone` once it is deleted. A page follows this rather than\nreading the collection again on a timer.",
        "operationId": "collection_events",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title",
            "description": "What the person called it; empty and the studio names it from its sources",
            "default": ""
          }
        },
        "type": "object",
        "title": "CollectionCreate"
      },
      "CollectionDetail": {
        "properties": {
          "collection": {
            "$ref": "#/components/schemas/CollectionSummary"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/SessionSummary"
            },
            "type": "array",
            "title": "Outputs",
            "description": "Decks and audio overviews, newest first"
          }
        },
        "type": "object",
        "required": [
          "collection",
          "outputs"
        ],
        "title": "CollectionDetail"
      },
      "CollectionPatch": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Empty hands the naming back to the studio"
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned"
          }
        },
        "type": "object",
        "title": "CollectionPatch"
      },
      "CollectionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "display_title": {
            "type": "string",
            "title": "Display Title",
            "description": "Its name as it is shown: the title, or `Untitled collection` while it has none"
          },
          "title_auto": {
            "type": "boolean",
            "title": "Title Auto",
            "description": "The studio named it; a person's title is never replaced"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The last time a source or an output changed"
          },
          "cover_version": {
            "type": "string",
            "title": "Cover Version",
            "description": "The cover drawn now: `f-\u2026` until one is designed, `g-\u2026` after. Load the cover with it as `v`, so a new one is fetched when it changes"
          },
          "sources": {
            "type": "integer",
            "title": "Sources"
          },
          "decks": {
            "type": "integer",
            "title": "Decks",
            "description": "Narrated slide decks, any state"
          },
          "audios": {
            "type": "integer",
            "title": "Audios",
            "description": "Audio overviews, any state"
          },
          "maps": {
            "type": "integer",
            "title": "Maps"
          },
          "notes": {
            "type": "integer",
            "title": "Notes"
          },
          "preparing": {
            "type": "integer",
            "title": "Preparing",
            "description": "Decks and audio overviews still being made"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "reused_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reused From",
            "description": "The share this collection was copied from by a reuse; null when none. The share may since have been removed"
          },
          "shared": {
            "type": "boolean",
            "title": "Shared",
            "description": "It has a share, so everyone on the studio sees it in Discover"
          },
          "reuses": {
            "type": "integer",
            "title": "Reuses",
            "description": "How many times its share was reused; 0 when not shared"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "description": "A copy whose share did not allow edits: it can be read, asked, pinned and deleted, but nothing in it changes and it cannot be shared"
          },
          "reused_from_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reused From Title",
            "description": "The name of the collection it was reused from, as it is shown now; null when it was not reused or that share is gone"
          },
          "busy": {
            "type": "boolean",
            "title": "Busy",
            "description": "Something in it is still being made: an output, its name or its cover. Follow /api/collections/{cid}/events while it is"
          },
          "auto_named": {
            "type": "boolean",
            "title": "Auto Named",
            "description": "The studio names it from its sources: automatic naming is on in Settings and nobody has given it a name"
          },
          "name_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Note",
            "description": "What the studio is doing with its name, in words: `Naming it from its sources\u2026` or `Named from its sources`; null when the name is a person's"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "display_title",
          "title_auto",
          "pinned",
          "created_at",
          "updated_at",
          "cover_version",
          "sources",
          "decks",
          "audios",
          "maps",
          "notes",
          "preparing",
          "failed",
          "reused_from",
          "shared",
          "reuses",
          "read_only",
          "reused_from_title",
          "busy",
          "auto_named",
          "name_note"
        ],
        "title": "CollectionSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}